Add fee structure and classes for Bitfinex
First commit only contains an implementation for Bitfinex.  The concept is that
there's the full fee structure for the exchange, which can be queried through
IExchange's "tradingFee" method, requiring a trading volume as well.  There's
also our individual fee for trading a pair using IAccount's "currentTradingFee"
method, which takes in less information and live queries the Exchange for your
particular fee.